+++ src/server/svgToJS.ts
@@ -1,6 +1,12 @@
 const attributesRE = /\s([a-zA-Z0-9-:]+)=("[^"]*")/gu;
 
-export const svgToJS = (svg: string, jsxImport: string, jsxFn: string) => {
+export const svgToJS = (
+  svg: string,
+  jsxImport: string,
+  jsxFn: string,
+  forwardRefImport: string,
+  forwardRefFn: string,
+) => {
   const index = svg.indexOf(">");
   const content = svg
     .slice(index + 1, svg.indexOf("</svg>"))
@@ -8,11 +14,12 @@ export const svgToJS = (svg: string, jsxImport: string, jsxFn: string) => {
     .replace(/\s+/g, " ");
   let attributes = "";
   for (const match of svg.slice(0, index).matchAll(attributesRE)) {
-    attributes += `  ${transformKey(match[1])}: ${match[2]},\n`;
+    attributes += `    ${transformKey(match[1])}: ${match[2]},\n`;
   }
   return `${jsxImport}
-export default ((props) => ${jsxFn}("svg", {
-${attributes}
+${forwardRefImport}  
+export default ${forwardRefFn}((props, ref) => ${jsxFn}("svg", {
+${attributes}    ref,
     ...props,
     dangerouslySetInnerHTML: { __html: '${content}' }
   })