+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
-## Unreleased
+## 0.4.0
 
 - Add SWC version to cache key for SWC cache
 - Use config.build.target (if defined) for bundling dependencies in development
+- Bump downwind to v0.3
+- Use forwardRef for SVG component
+- Fix: Make devtools update works when URL is not root
 
 ## 0.3.1